fix: sync the start with sessionStorage.
When a new session starts, the page is redirected to auth. But when it is re-directed back with a token, all the search params are removed(our backend does not support state).
So I sync the start to sessionStorage, when the page is redirect to auth, the searchParam is saved to sessionStorage, then it can be recovered when receiving the token.